+++ src/components/MembersProgressTable/membersProgressTable.scss
@@ -1,8 +1,8 @@
-.admin-btns {
+.user-btns {
   display: grid;
   grid-gap: 5px;
   grid-template-columns: min-content min-content;
-  .custom-btn {
-    justify-self: self-start;
+  button {
+    margin: 0;
   }
 }
