+++ src/index.scss
@@ -10,6 +10,13 @@ body {
   font-size: 16px;
 }
 
+td,
+th {
+  vertical-align: middle !important;
+  text-align: center !important;
+  width: fit-content !important;
+}
+
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
 }
