+++ src/pages/TasksManagePage.js
@@ -9,7 +9,7 @@ import { defaultTaskData } from '../utils/defaultInputsData';
 import { tasksInputs } from '../utils/inputs';
 import inputsChangeHandler from '../utils/inputsChangeHandler';
 import validation from '../utils/validation';
-import ModalContent from '../UI/Modal';
+import ModalContent from '../UI/ModalContent';
 import DataModal from '../components/DataModal';
 import FormModal from '../components/FormModal';
 import { stringToDate, dateToString } from '../utils/convertDate';
